requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.6.2.post1
arrow==1.3.0
asn1==2.7.1
asttokens==3.0.0
blessed==1.20.0
blinker==1.9.0
bpylist2==4.1.1
cached-property==2.0.1
case-insensitive-dictionary==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
construct==2.10.70
cryptography==44.0.0
daemonize==2.5.0
decorator==5.1.1
developer_disk_image==0.0.4
dnspython==2.7.0
editor==1.6.6
email_validator==2.2.0
enum-compat==0.0.3
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.6
Flask==3.1.0
Flask-Cors==5.0.0
gpxpy==1.5.0
h11==0.14.0
hexdump==3.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ifaddr==0.2.0
imageio==2.35.1
iniconfig==2.0.0
inquirer3==0.6.1
ipsw-parser==1.3.9
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jinxed==1.3.0
la-panic==0.5.0
lzfse==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpi4py==4.0.1
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.1.1
opack==0.1.0
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
parameter-decorators==0.0.2
parso==0.8.4
pillow==10.4.0
pluggy==1.5.0
plumbum==1.9.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
pycrashreport==1.2.5
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pygnuutils==0.1.1
pyimg4==0.8.6
pykdebugparser==1.2.4
pylzss==0.3.4
pymobiledevice3==4.16.3
pyreadline3==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-pcapng==2.1.1
pytun-pmd3==2.0.9
pyusb==1.2.1
pywin32==308
PyYAML==6.0.2
qh3==1.2.1
readchar==4.2.1
remotezip==0.12.3
remotezip2==0.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.11.3
runs==1.2.2
scipy==1.14.1
shellingham==1.5.4
SideJITServer==1.3.6
six==1.17.0
sk-video==1.1.10
sniffio==1.3.1
srptools==1.0.1
sslpsk-pmd3==1.0.3
stack-data==0.6.3
starlette==0.41.3
termcolor==2.5.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
xmod==1.8.1
xonsh==0.18.4
zeroconf==0.132.2
